feat(bsp_graphics_lvgl): Created new component for BSP graphics (LVGL) and removed LVGL from BSP
