+++ bsp/m5stack_tab5/idf_component.yml
@@ -1,5 +1,5 @@
 
-version: "1.2.0~1"
+version: "2.0.0"
 description: Board Support Package (BSP) for M5Stack Tab5
 url: https://github.com/espressif/esp-bsp/tree/master/bsp/m5stack_tab5
 
@@ -27,11 +27,6 @@ dependencies:
     public: true
     override_path: "../../components/io_expander/esp_io_expander_pi4ioe5v6408"
 
-  espressif/esp_lvgl_port:
-    version: "^2"
-    public: true
-    override_path: "../../components/esp_lvgl_port"
-
   espressif/esp_lcd_st7123:
     version: "*"
     public: false
@@ -43,11 +38,11 @@ dependencies:
 
   espressif/esp_lcd_touch_st7123:
     version: "*"
-    public: false
+    public: true
 
   espressif/esp_lcd_touch_gt911:
     version: "*"
-    public: false
+    public: true
     # Here cannot be override_path, because we are using two touch comonents.
     # The ST7123 is not owned by BSP and loading another version of esp_lcd_touch
