+++ bsp/m5stack_tab5/include/bsp/m5stack_tab5.h
@@ -24,11 +24,6 @@
 #include "esp_io_expander.h"
 #include "iot_sensor_hub.h"
 
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-#include "lvgl.h"
-#include "esp_lvgl_port.h"
-#endif // BSP_CONFIG_NO_GRAPHIC_LIB == 0
 /**************************************************************************************************
  *  BSP Board Name
  **************************************************************************************************/
@@ -402,118 +397,6 @@ esp_err_t bsp_sdcard_sdmmc_mount(bsp_sdcard_cfg_t *cfg);
 esp_err_t bsp_sdcard_sdspi_mount(bsp_sdcard_cfg_t *cfg);
 
 /** @} */ // end of storage
-
-/** \addtogroup g04_display
- *  @{
- */
-
-/**************************************************************************************************
- *
- * LCD interface
- *
- * LVGL is used as graphics library. LVGL is NOT thread safe, therefore the user must take LVGL mutex
- * by calling bsp_display_lock() before calling any LVGL API (lv_...) and then give the mutex with
- * bsp_display_unlock().
- *
- * Display's backlight must be enabled explicitly by calling bsp_display_backlight_on()
- **************************************************************************************************/
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
-/**
- * @brief BSP display (LVGL) configuration structure
- */
-typedef struct {
-    lvgl_port_cfg_t lvgl_port_cfg;  /*!< LVGL port configuration */
-    uint32_t        buffer_size;    /*!< Size of the buffer for the screen in pixels */
-    bool            double_buffer;  /*!< True, if should be allocated two buffers */
-    struct {
-        unsigned int buff_dma: 1;    /*!< Allocated LVGL buffer will be DMA capable */
-        unsigned int buff_spiram: 1; /*!< Allocated LVGL buffer will be in PSRAM */
-        unsigned int sw_rotate: 1;   /*!< Use software rotation (slower), The feature is unavailable under avoid-tear mode */
-    } flags;
-} bsp_display_cfg_t;
-
-/**
- * @brief Initialize display
- *
- * This function initializes SPI, display controller and starts LVGL handling task.
- * LCD backlight must be enabled separately by calling bsp_display_brightness_set()
- *
- * @return Pointer to LVGL display or NULL when error occurred
- */
-lv_display_t *bsp_display_start(void);
-
-/**
- * @brief Initialize display
- *
- * This function initializes SPI, display controller and starts LVGL handling task.
- * LCD backlight must be enabled separately by calling bsp_display_brightness_set()
- *
- * @param cfg display configuration
- *
- * @return Pointer to LVGL display or NULL when error occurred
- */
-lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg);
-
-/**
- * @brief Get pointer to input device (touch, buttons, ...)
- *
- * @note The LVGL input device is initialized in bsp_display_start() function.
- *
- * @return Pointer to LVGL input device or NULL when not initialized
- */
-lv_indev_t *bsp_display_get_input_dev(void);
-
-/**
- * @brief Take LVGL mutex
- *
- * @param timeout_ms Timeout in [ms]. 0 will block indefinitely.
- * @return true  Mutex was taken
- * @return false Mutex was NOT taken
- */
-bool bsp_display_lock(uint32_t timeout_ms);
-
-/**
- * @brief Give LVGL mutex
- *
- */
-void bsp_display_unlock(void);
-
-/**
- * @brief Set display enter sleep mode
- *
- * All the display (LCD, backlight, touch) will enter sleep mode.
- *
- * @return
- *      - ESP_OK on success
- *      - ESP_ERR_NOT_SUPPORTED if this function is not supported by the panel
- */
-esp_err_t bsp_display_enter_sleep(void);
-
-/**
- * @brief Set display exit sleep mode
- *
- * All the display (LCD, backlight, touch) will exit sleep mode.
- *
- * @return
- *      - ESP_OK on success
- *      - ESP_ERR_NOT_SUPPORTED if this function is not supported by the panel
- */
-esp_err_t bsp_display_exit_sleep(void);
-
-/**
- * @brief Rotate screen
- *
- * Display must be already initialized by calling bsp_display_start()
- *
- * @param[in] disp Pointer to LVGL display
- * @param[in] rotation Angle of the display rotation
- */
-void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation);
-#endif // BSP_CONFIG_NO_GRAPHIC_LIB == 0
-
-/** @} */ // end of display
 
 /** \addtogroup g12_camera
  *  @{
