+++ bsp/m5stack_tab5/src/bsp_display.c
@@ -28,14 +28,9 @@
 
 static const char *TAG = "M5Stack Tab5";
 
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_display_t *disp;
-static lv_indev_t *disp_indev_touch = NULL;
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
 static bsp_lcd_handles_t disp_handles;
+static esp_lcd_touch_handle_t tp;
 static esp_ldo_channel_handle_t disp_phy_pwr_chan = NULL;
-static esp_lcd_touch_handle_t tp;   // LCD touch handle
 
 // Bit number used to represent command and parameter
 #define LCD_CMD_BITS           8
@@ -323,69 +318,6 @@ void bsp_display_delete(void)
     bsp_display_brightness_deinit();
 }
 
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_display_t *bsp_display_lcd_init(const bsp_display_cfg_t *cfg)
-{
-    assert(cfg != NULL);
-    esp_lcd_panel_io_handle_t io_handle = NULL;
-    const bsp_display_config_t bsp_disp_cfg = {
-        .dsi_bus = {
-            .phy_clk_src = 0, // let the driver to choose the default clock source
-            .lane_bit_rate_mbps = BSP_LCD_MIPI_DSI_LANE_BITRATE_MBPS,
-        }
-    };
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_new(&bsp_disp_cfg, &disp_handles.panel, &io_handle));
-
-    esp_lcd_panel_disp_on_off(disp_handles.panel, true);
-
-    /* Add LCD screen */
-    ESP_LOGD(TAG, "Add LCD screen");
-    const lvgl_port_display_cfg_t disp_cfg = {
-        .io_handle = io_handle,
-        .panel_handle = disp_handles.panel,
-        .buffer_size = cfg->buffer_size,
-        .double_buffer = cfg->double_buffer,
-        .hres = BSP_LCD_H_RES,
-        .vres = BSP_LCD_V_RES,
-        .monochrome = false,
-        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
-        .rotation = {
-            .swap_xy = false,
-            .mirror_x = false,
-            .mirror_y = false,
-        },
-        .flags = {
-            .buff_dma = cfg->flags.buff_dma,
-            .buff_spiram = cfg->flags.buff_spiram,
-#if LVGL_VERSION_MAJOR >= 9
-            .swap_bytes = (BSP_LCD_BIGENDIAN ? true : false),
-#endif
-#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
-            .sw_rotate = false,                /* Avoid tearing is not supported for SW rotation */
-#else
-            .sw_rotate = cfg->flags.sw_rotate,
-#endif
-#if CONFIG_BSP_DISPLAY_LVGL_FULL_REFRESH
-            .full_refresh = true,
-#elif CONFIG_BSP_DISPLAY_LVGL_DIRECT_MODE
-            .direct_mode = true,
-#endif
-        }
-    };
-    const lvgl_port_display_dsi_cfg_t dpi_cfg = {
-        .flags = {
-#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
-            .avoid_tearing = true,
-#else
-            .avoid_tearing = false,
-#endif
-        }
-    };
-
-    return lvgl_port_add_disp_dsi(&disp_cfg, &dpi_cfg);
-}
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
 esp_err_t bsp_touch_new(const bsp_touch_config_t *config, esp_lcd_touch_handle_t *ret_touch)
 {
     /* Initilize I2C */
@@ -442,22 +374,9 @@ esp_err_t bsp_touch_new(const bsp_touch_config_t *config, esp_lcd_touch_handle_t
                             TAG, "New touch driver initialization failed");
     }
 
+    tp = *ret_touch;
+
     return ESP_OK;
-}
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-static lv_indev_t *bsp_display_indev_touch_init(lv_display_t *disp)
-{
-    BSP_ERROR_CHECK_RETURN_NULL(bsp_touch_new(NULL, &tp));
-    assert(tp);
-
-    /* Add touch input (for selected screen) */
-    const lvgl_port_touch_cfg_t touch_cfg = {
-        .disp = disp,
-        .handle = tp,
-    };
-
-    return lvgl_port_add_touch(&touch_cfg);
 }
 
 static esp_err_t bsp_touch_enter_sleep(void)
@@ -470,57 +389,6 @@ static esp_err_t bsp_touch_exit_sleep(void)
 {
     assert(tp);
     return esp_lcd_touch_exit_sleep(tp);
-}
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-
-#if (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
-lv_display_t *bsp_display_start(void)
-{
-    bsp_display_cfg_t cfg = {
-        .lvgl_port_cfg = ESP_LVGL_PORT_INIT_CONFIG(),
-        .buffer_size = BSP_LCD_H_RES * CONFIG_BSP_LCD_DRAW_BUF_HEIGHT,
-#if CONFIG_BSP_LCD_DRAW_BUF_DOUBLE
-        .double_buffer = 1,
-#else
-        .double_buffer = 0,
-#endif
-        .flags = {
-            .buff_dma = true,
-            .buff_spiram = false,
-            .sw_rotate = true,
-        }
-    };
-    return bsp_display_start_with_config(&cfg);
-}
-
-lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg)
-{
-    assert(cfg != NULL);
-    BSP_ERROR_CHECK_RETURN_NULL(lvgl_port_init(&cfg->lvgl_port_cfg));
-    BSP_NULL_CHECK(disp = bsp_display_lcd_init(cfg), NULL);
-    BSP_NULL_CHECK(disp_indev_touch = bsp_display_indev_touch_init(disp), NULL);
-
-    return disp;
-}
-
-lv_indev_t *bsp_display_get_input_dev(void)
-{
-    return disp_indev_touch;
-}
-
-void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation)
-{
-    lv_disp_set_rotation(disp, rotation);
-}
-
-bool bsp_display_lock(uint32_t timeout_ms)
-{
-    return lvgl_port_lock(timeout_ms);
-}
-
-void bsp_display_unlock(void)
-{
-    lvgl_port_unlock();
 }
 
 static esp_err_t bsp_lcd_enter_sleep(void)
@@ -550,4 +418,3 @@ esp_err_t bsp_display_exit_sleep(void)
     BSP_ERROR_CHECK_RETURN_ERR(bsp_touch_exit_sleep());
     return ESP_OK;
 }
-#endif // (BSP_CONFIG_NO_GRAPHIC_LIB == 0)
