+++ components/bsp_graphics/README.md
@@ -1,0 +1,3 @@
+# BSP Graphics
+
+This folder contains graphics components (LVGL, ...). All these components can be found in [IDF Component Registry](https://components.espressif.com/).
