+++ components/bsp_graphics/bsp_graphics_lvgl/src/bsp_graphics_lvgl.c
@@ -1,0 +1,159 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+
+/*******************************************************************************
+* Includes
+*******************************************************************************/
+#include "esp_err.h"
+#include "esp_log.h"
+#include "esp_check.h"
+#include "bsp_err_check.h"
+#include "esp_lcd_panel_io.h"
+#include "bsp/display.h"
+#include "bsp/touch.h"
+#include "bsp/esp-bsp.h"
+#include "esp_bsp_graphics.h"
+
+/*******************************************************************************
+* Local variables
+*******************************************************************************/
+static const char *TAG = "BSP Graphics";
+static lv_display_t *disp;
+static lv_indev_t *disp_indev_touch = NULL;
+
+/*******************************************************************************
+* Private functions
+*******************************************************************************/
+static lv_display_t *bsp_display_lcd_init(const bsp_display_cfg_t *cfg)
+{
+    bsp_lcd_handles_t disp_handles;
+    assert(cfg != NULL);
+    esp_lcd_panel_io_handle_t io_handle = NULL;
+    const bsp_display_config_t bsp_disp_cfg = {
+        .dsi_bus = {
+            .phy_clk_src = 0, // let the driver to choose the default clock source
+            .lane_bit_rate_mbps = BSP_LCD_MIPI_DSI_LANE_BITRATE_MBPS,
+        }
+    };
+    BSP_ERROR_CHECK_RETURN_NULL(bsp_display_new(&bsp_disp_cfg, &disp_handles.panel, &io_handle));
+
+    esp_lcd_panel_disp_on_off(disp_handles.panel, true);
+
+    /* Add LCD screen */
+    ESP_LOGD(TAG, "Add LCD screen");
+    const lvgl_port_display_cfg_t disp_cfg = {
+        .io_handle = io_handle,
+        .panel_handle = disp_handles.panel,
+        .buffer_size = cfg->buffer_size,
+        .double_buffer = cfg->double_buffer,
+        .hres = BSP_LCD_H_RES,
+        .vres = BSP_LCD_V_RES,
+        .monochrome = false,
+        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
+        .rotation = {
+            .swap_xy = false,
+            .mirror_x = false,
+            .mirror_y = false,
+        },
+        .flags = {
+            .buff_dma = cfg->flags.buff_dma,
+            .buff_spiram = cfg->flags.buff_spiram,
+#if LVGL_VERSION_MAJOR >= 9
+            .swap_bytes = (BSP_LCD_BIGENDIAN ? true : false),
+#endif
+#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
+            .sw_rotate = false,                /* Avoid tearing is not supported for SW rotation */
+#else
+            .sw_rotate = cfg->flags.sw_rotate,
+#endif
+#if CONFIG_BSP_DISPLAY_LVGL_FULL_REFRESH
+            .full_refresh = true,
+#elif CONFIG_BSP_DISPLAY_LVGL_DIRECT_MODE
+            .direct_mode = true,
+#endif
+        }
+    };
+    const lvgl_port_display_dsi_cfg_t dpi_cfg = {
+        .flags = {
+#if CONFIG_BSP_DISPLAY_LVGL_AVOID_TEAR
+            .avoid_tearing = true,
+#else
+            .avoid_tearing = false,
+#endif
+        }
+    };
+
+    return lvgl_port_add_disp_dsi(&disp_cfg, &dpi_cfg);
+}
+
+static lv_indev_t *bsp_display_indev_touch_init(lv_display_t *disp)
+{
+    esp_lcd_touch_handle_t tp;
+    BSP_ERROR_CHECK_RETURN_NULL(bsp_touch_new(NULL, &tp));
+    assert(tp);
+
+    /* Add touch input (for selected screen) */
+    const lvgl_port_touch_cfg_t touch_cfg = {
+        .disp = disp,
+        .handle = tp,
+    };
+
+    return lvgl_port_add_touch(&touch_cfg);
+}
+
+/*******************************************************************************
+* Public API functions
+*******************************************************************************/
+
+lv_display_t *bsp_display_start(void)
+{
+    bsp_display_cfg_t cfg = {
+        .lvgl_port_cfg = ESP_LVGL_PORT_INIT_CONFIG(),
+        .buffer_size = BSP_LCD_H_RES * CONFIG_BSP_LCD_DRAW_BUF_HEIGHT,
+#if CONFIG_BSP_LCD_DRAW_BUF_DOUBLE
+        .double_buffer = 1,
+#else
+        .double_buffer = 0,
+#endif
+        .flags = {
+            .buff_dma = true,
+            .buff_spiram = false,
+            .sw_rotate = true,
+        }
+    };
+    return bsp_display_start_with_config(&cfg);
+}
+
+lv_display_t *bsp_display_start_with_config(const bsp_display_cfg_t *cfg)
+{
+    assert(cfg != NULL);
+    BSP_ERROR_CHECK_RETURN_NULL(lvgl_port_init(&cfg->lvgl_port_cfg));
+    BSP_NULL_CHECK(disp = bsp_display_lcd_init(cfg), NULL);
+    BSP_NULL_CHECK(disp_indev_touch = bsp_display_indev_touch_init(disp), NULL);
+
+    return disp;
+}
+
+lv_indev_t *bsp_display_get_input_dev(void)
+{
+    return disp_indev_touch;
+}
+
+void bsp_display_rotate(lv_display_t *disp, lv_disp_rotation_t rotation)
+{
+    lv_disp_set_rotation(disp, rotation);
+}
+
+bool bsp_display_lock(uint32_t timeout_ms)
+{
+    return lvgl_port_lock(timeout_ms);
+}
+
+void bsp_display_unlock(void)
+{
+    lvgl_port_unlock();
+}
