+++ examples/display_lvgl_demos/main/main.c
@@ -17,6 +17,9 @@
 
 #include "lv_demos.h"
 #include "bsp/esp-bsp.h"
+#ifdef BSP_GRAPHICS
+#include "esp_bsp_graphics.h"
+#endif
 
 static char *TAG = "app_main";
